# perf: subgroup check