--- a/ecc/bls12-377/g2.go
+++ b/ecc/bls12-377/g2.go
@@ -494,7 +494,7 @@ func (p *G2Jac) IsOnCurve() bool {
 func (p *G2Jac) IsInSubGroup() bool {
 	var res, tmp G2Jac
 	tmp.psi(p)
-	res.ScalarMultiplication(p, &xGen).
+	res.mulWindowed(p, &xGen).
 		SubAssign(&tmp)
 
 	return res.IsOnCurve() && res.Z.IsZero()
@@ -638,8 +638,8 @@ func (p *G2Affine) ClearCofactor(a *G2Affine) *G2Affine {
 func (p *G2Jac) ClearCofactor(q *G2Jac) *G2Jac {
 	// https://eprint.iacr.org/2017/419.pdf, 4.1
 	var xg, xxg, res, t G2Jac
-	xg.ScalarMultiplication(q, &xGen)
-	xxg.ScalarMultiplication(&xg, &xGen)
+	xg.mulWindowed(q, &xGen)
+	xxg.mulWindowed(&xg, &xGen)
 
 	res.Set(&xxg).
 		SubAssign(&xg).