# Description

use `mulWindowed` instead of `ScalarMultiplication` is G1 and G2 subgroup membership and cofactor clearing methods. 
1- The point is still purported to be in the prime subgroup where GLV applies
2- multiplying by a small constant is faster with `mulWindowed` 

N.B.: multiplying e.g. twice by `x0` is faster than multiplying by `x0^2` because it has smaller Hamming weight. E.g. for BLS12-381, `mul-by-x0` is 64 dbl + 6 mul while `mul-by-x0^2` is 128 dbl + 17 mul. There is still room for improvement if needed by specialising `mul-by-x0` using a short addition chain.

## Type of change

Please delete options that are not relevant.

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update

# How has this been tested?

Same tests.

# How has this been benchmarked?

- BLS12-38
```
benchmark                        old ns/op     new ns/op     delta
BenchmarkG1JacIsInSubGroup-2     59891         46740         -21.96%
```
```
benchmark                        old ns/op     new ns/op     delta
BenchmarkG2JacIsInSubGroup-2     74121         55710          -24.84%
```

- BN254
```
benchmark                        old ns/op     new ns/op     delta
BenchmarkG2JacIsInSubGroup-2     58285         48690        -16.46%
```

- BW6-761
```
benchmark                        old ns/op     new ns/op     delta
BenchmarkG1JacIsInSubGroup-2     410121        322322        -21.41%
```
```
benchmark                        old ns/op     new ns/op     delta
BenchmarkG2JacIsInSubGroup-2     410955        322196        -21.60%
```

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] I did not modify files generated from templates
- [x] `golangci-lint` does not output errors locally
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream modules